started styling navigation bar
